added modifications to deselect difficulty and link 3 first activities to their test pages
