+++ .idea/vcs.xml
@@ -2,5 +2,7 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/Learning activities" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/Master" vcs="Git" />
   </component>
 </project>
