+++ app/src/main/java/com/example/projectdva232v1/MainActivity.kt
@@ -5,9 +5,11 @@ import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.drawerlayout.widget.DrawerLayout
+import androidx.fragment.app.commit
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.NavigationUI
+import com.example.projectdva232v1.ui.HomePage.homePageFragment
 import com.google.android.material.navigation.NavigationView
 
 
@@ -62,6 +64,7 @@ class MainActivity : AppCompatActivity() {
             else -> super.onOptionsItemSelected(item)
         }
     }
+
 
 
 }
