+++ app/src/main/java/com/example/projectdva232v1/ui/HomePage/ActivityItem.kt
@@ -1,3 +1,3 @@
-package com.example.projectdva232v1.ui.homePage
+package com.example.projectdva232v1.ui.HomePage
 
 data class ActivityItem(var activityName: String, var imageResource:Int){}
