+++ app/src/main/java/com/example/projectdva232v1/ui/HomePage/DifficultyLevelItem.kt
@@ -1,3 +1,3 @@
-package com.example.projectdva232v1.ui.homePage
+package com.example.projectdva232v1.ui.HomePage
 
 data class DifficultyLevelItem(var difficultyName: String){}
