+++ app/src/main/java/com/example/projectdva232v1/ui/HomePage/RecyclerAdapterActivity.kt
@@ -1,30 +1,25 @@
-package com.example.projectdva232v1.ui.ReadingActivity
+package com.example.projectdva232v1.ui.HomePage
 
-import android.graphics.Color
-import android.text.SpannableString
-import android.text.style.ForegroundColorSpan
-import android.text.style.RelativeSizeSpan
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.projectdva232v1.R
-import com.example.projectdva232v1.ui.homePage.DifficultyLevelItem
 
 
-class RecyclerAdapterReading(
-        private val activityList: List<DifficultyLevelItem>,
-        private val listener: OnItemClickListener,
-        var score: Int = 0
-
-) : RecyclerView.Adapter<RecyclerAdapterReading.ViewHolder>() {
+class RecyclerAdapterActivity(
+        private val activityList: List<ActivityItem>,
+        private val listener: OnItemClickListener
+) : RecyclerView.Adapter<RecyclerAdapterActivity.ViewHolder>() {
 
     // Called for every ViewHolder item that is to be created
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         // Build view from XML file
-        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.recyclerview_readingitem,
+        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.recyclerview_activityitem,
                 parent, false)
+
         return ViewHolder(itemView)
     }
 
@@ -36,20 +31,24 @@ class RecyclerAdapterReading(
     // Takes ViewHolder and fills it with the new data to be displayed
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val currentItem = activityList[position]
-        //TODO: get best score and difficulty from previous attempts & page
-        val s = "Test ${position+1}\nBest score: $score\nLevel: ${currentItem.difficultyName}"
-        val span = SpannableString(s)
-        span.setSpan(RelativeSizeSpan(1.5f),0,span.indexOf("\n"),0)
-        span.setSpan(ForegroundColorSpan(Color.GRAY),span.indexOf("\n"),span.length,0)
-        holder.textActivity.text = span
-        holder.textImageView.text = (position + 1).toString()
+        holder.imageView.setImageResource(currentItem.imageResource)
+        holder.textActivity.text = currentItem.activityName
+
+        //sets the background color of the current item
+        when(position){
+            0 -> holder.imageView.setBackgroundResource(R.color.listening)
+            1 -> holder.imageView.setBackgroundResource(R.color.vocabulary)
+            2 -> holder.imageView.setBackgroundResource(R.color.writing)
+            3 -> holder.imageView.setBackgroundResource(R.color.speaking)
+            4 -> holder.imageView.setBackgroundResource(R.color.reading)
+        }
     }
 
     // Class with the data that is displayed for each ActivityItem
     inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView),
             View.OnClickListener {
-        val textImageView: TextView = itemView.findViewById(R.id.reading_circle_tv)
-        val textActivity: TextView = itemView.findViewById(R.id.reading_tv)
+        val imageView: ImageView = itemView.findViewById(R.id.activity_icon)
+        val textActivity: TextView = itemView.findViewById(R.id.activity_tv)
 
         init {
             itemView.setOnClickListener(this)
@@ -65,5 +64,6 @@ class RecyclerAdapterReading(
 
     interface OnItemClickListener {
         fun onItemClick(position: Int)
+        fun onItemClicked(difficulty:String)
     }
 }
