+++ app/src/main/java/com/example/projectdva232v1/ui/Listening/ListeningModel.kt
@@ -1,10 +1,10 @@
-package com.example.projectdva232v1.ui.SpeakingActivity
+package com.example.projectdva232v1.ui.Listening
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SpeakingModel : ViewModel() {
+class ListeningModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is reading Fragment"
