+++ app/src/main/java/com/example/projectdva232v1/ui/Reading/RecyclerAdapterReading.kt
@@ -1,27 +1,30 @@
-package com.example.projectdva232v1.ui.homePage
+package com.example.projectdva232v1.ui.Reading
 
+import android.graphics.Color
+import android.text.SpannableString
+import android.text.style.ForegroundColorSpan
+import android.text.style.RelativeSizeSpan
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TableRow
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.example.projectdva232v1.MainActivity
 import com.example.projectdva232v1.R
+import com.example.projectdva232v1.ui.HomePage.DifficultyLevelItem
 
 
-class RecyclerAdapterActivity(
-        private val activityList: List<ActivityItem>,
-        private val listener: OnItemClickListener
-) : RecyclerView.Adapter<RecyclerAdapterActivity.ViewHolder>() {
+class RecyclerAdapterReading(
+        private val activityList: List<DifficultyLevelItem>,
+        private val listener: OnItemClickListener,
+        var score: Int = 0
+
+) : RecyclerView.Adapter<RecyclerAdapterReading.ViewHolder>() {
 
     // Called for every ViewHolder item that is to be created
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         // Build view from XML file
-        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.recyclerview_activityitem,
+        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.recyclerview_readingitem,
                 parent, false)
-
         return ViewHolder(itemView)
     }
 
@@ -33,24 +36,20 @@ class RecyclerAdapterActivity(
     // Takes ViewHolder and fills it with the new data to be displayed
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val currentItem = activityList[position]
-        holder.imageView.setImageResource(currentItem.imageResource)
-        holder.textActivity.text = currentItem.activityName
-
-        //sets the background color of the current item
-        when(position){
-            0 -> holder.imageView.setBackgroundResource(R.color.listening)
-            1 -> holder.imageView.setBackgroundResource(R.color.vocabulary)
-            2 -> holder.imageView.setBackgroundResource(R.color.writing)
-            3 -> holder.imageView.setBackgroundResource(R.color.speaking)
-            4 -> holder.imageView.setBackgroundResource(R.color.reading)
-        }
+        //TODO: get best score and difficulty from previous attempts & page
+        val s = "Test ${position+1}\nBest score: $score\nLevel: ${currentItem.difficultyName}"
+        val span = SpannableString(s)
+        span.setSpan(RelativeSizeSpan(1.5f),0,span.indexOf("\n"),0)
+        span.setSpan(ForegroundColorSpan(Color.GRAY),span.indexOf("\n"),span.length,0)
+        holder.textActivity.text = span
+        holder.textImageView.text = (position + 1).toString()
     }
 
     // Class with the data that is displayed for each ActivityItem
     inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView),
             View.OnClickListener {
-        val imageView: ImageView = itemView.findViewById(R.id.activity_icon)
-        val textActivity: TextView = itemView.findViewById(R.id.activity_tv)
+        val textImageView: TextView = itemView.findViewById(R.id.reading_circle_tv)
+        val textActivity: TextView = itemView.findViewById(R.id.reading_tv)
 
         init {
             itemView.setOnClickListener(this)
@@ -66,6 +65,5 @@ class RecyclerAdapterActivity(
 
     interface OnItemClickListener {
         fun onItemClick(position: Int)
-        fun onItemClicked(difficulty:String)
     }
 }
