+++ app/src/main/java/com/example/projectdva232v1/ui/Speaking/SpeakingFragment.kt
@@ -1,4 +1,4 @@
-package com.example.projectdva232v1.ui.VocabularyActivity
+package com.example.projectdva232v1.ui.Speaking
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,17 +7,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import androidx.navigation.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.projectdva232v1.R
-import com.example.projectdva232v1.ui.homePage.DifficultyLevelItem
-import com.example.projectdva232v1.ui.learning_activities.VocabularyActivity
+import com.example.projectdva232v1.ui.HomePage.DifficultyLevelItem
+import com.example.projectdva232v1.ui.learning_activities.SpeakingActivity
 
 
-class VocabularyFragment : Fragment(), RecyclerAdapterVocabulary.OnItemClickListener {
+class SpeakingFragment : Fragment(), RecyclerAdapterSpeaking.OnItemClickListener {
 
-    private lateinit var vocabularyModel: VocabularyModel
+    private lateinit var speakingModel: SpeakingModel
     var activityList = ArrayList<DifficultyLevelItem>()
     lateinit var difficultySelected: String
 
@@ -27,14 +26,15 @@ class VocabularyFragment : Fragment(), RecyclerAdapterVocabulary.OnItemClickList
         savedInstanceState: Bundle?
     ): View? {
 
-        difficultySelected = VocabularyFragmentArgs.fromBundle(requireArguments()).VocabularyFragmentArgs
+        difficultySelected = SpeakingFragmentArgs.fromBundle(requireArguments()).SpeakingFragmentArgs
 
-        vocabularyModel =
-            ViewModelProvider(this).get(VocabularyModel::class.java)
+        speakingModel =
+            ViewModelProvider(this).get(SpeakingModel::class.java)
         val root = inflater.inflate(R.layout.fragment_reading_activity, container, false)
 
         //Since we don't have access to the API
         val activities = resources.getStringArray(R.array.difficulty_array)
+        activityList.clear()
         for (diff in activities) {
             if (diff == difficultySelected) {
                 val item = DifficultyLevelItem(diff)
@@ -43,19 +43,18 @@ class VocabularyFragment : Fragment(), RecyclerAdapterVocabulary.OnItemClickList
         }
         // Setup for the recycler view
         val rv = root.findViewById<RecyclerView>(R.id.recycler_view_activity_reading)
-        rv.adapter = RecyclerAdapterVocabulary(activityList, this)
+        rv.adapter = RecyclerAdapterSpeaking(activityList, this)
         rv.layoutManager = LinearLayoutManager(activity, LinearLayoutManager.VERTICAL, false)
         rv.setHasFixedSize(true)
         return root
     }
 
     override fun onItemClick(position: Int) {
-
         for (index in 0..activityList.size) {
             if (position == index) {
-                // view?.findNavController()?.navigate(R.id.action_nav_vocabulary_to_nav_example5)
+                // view?.findNavController()?.navigate(R.id.action_nav_speaking_to_nav_example5)
 
-                val intent = Intent(this.context, VocabularyActivity::class.java)
+                val intent = Intent(this.context, SpeakingActivity::class.java)
                 startActivity(intent)
             }
         }
