+++ app/src/main/java/com/example/projectdva232v1/ui/Speaking/SpeakingModel.kt
@@ -1,10 +1,10 @@
-package com.example.projectdva232v1.ui.ReadingActivity
+package com.example.projectdva232v1.ui.Speaking
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ReadingModel : ViewModel() {
+class SpeakingModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is reading Fragment"
