+++ app/src/main/java/com/example/projectdva232v1/ui/Vocabulary/VocabularyFragment.kt
@@ -1,4 +1,4 @@
-package com.example.projectdva232v1.ui.SpeakingActivity
+package com.example.projectdva232v1.ui.Vocabulary
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,17 +7,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import androidx.navigation.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.projectdva232v1.R
-import com.example.projectdva232v1.ui.homePage.DifficultyLevelItem
-import com.example.projectdva232v1.ui.learning_activities.SpeakingActivity
+import com.example.projectdva232v1.ui.HomePage.DifficultyLevelItem
+import com.example.projectdva232v1.ui.learning_activities.VocabularyActivity
 
 
-class SpeakingFragment : Fragment(), RecyclerAdapterSpeaking.OnItemClickListener {
+class VocabularyFragment : Fragment(), RecyclerAdapterVocabulary.OnItemClickListener {
 
-    private lateinit var speakingModel: SpeakingModel
+    private lateinit var vocabularyModel: VocabularyModel
     var activityList = ArrayList<DifficultyLevelItem>()
     lateinit var difficultySelected: String
 
@@ -27,14 +26,15 @@ class SpeakingFragment : Fragment(), RecyclerAdapterSpeaking.OnItemClickListener
         savedInstanceState: Bundle?
     ): View? {
 
-        difficultySelected = SpeakingFragmentArgs.fromBundle(requireArguments()).SpeakingFragmentArgs
+        difficultySelected = VocabularyFragmentArgs.fromBundle(requireArguments()).VocabularyFragmentArgs
 
-        speakingModel =
-            ViewModelProvider(this).get(SpeakingModel::class.java)
+        vocabularyModel =
+            ViewModelProvider(this).get(VocabularyModel::class.java)
         val root = inflater.inflate(R.layout.fragment_reading_activity, container, false)
 
         //Since we don't have access to the API
         val activities = resources.getStringArray(R.array.difficulty_array)
+        activityList.clear()
         for (diff in activities) {
             if (diff == difficultySelected) {
                 val item = DifficultyLevelItem(diff)
@@ -43,18 +43,18 @@ class SpeakingFragment : Fragment(), RecyclerAdapterSpeaking.OnItemClickListener
         }
         // Setup for the recycler view
         val rv = root.findViewById<RecyclerView>(R.id.recycler_view_activity_reading)
-        rv.adapter = RecyclerAdapterSpeaking(activityList, this)
+        rv.adapter = RecyclerAdapterVocabulary(activityList, this)
         rv.layoutManager = LinearLayoutManager(activity, LinearLayoutManager.VERTICAL, false)
         rv.setHasFixedSize(true)
         return root
     }
 
     override fun onItemClick(position: Int) {
+
         for (index in 0..activityList.size) {
             if (position == index) {
-                // view?.findNavController()?.navigate(R.id.action_nav_speaking_to_nav_example5)
 
-                val intent = Intent(this.context, SpeakingActivity::class.java)
+                val intent = Intent(this.context, VocabularyActivity::class.java)
                 startActivity(intent)
             }
         }
