+++ app/src/main/java/com/example/projectdva232v1/ui/Vocabulary/VocabularyModel.kt
@@ -1,14 +1,13 @@
-package com.example.projectdva232v1.ui.homePage
+package com.example.projectdva232v1.ui.Vocabulary
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomePageModel : ViewModel() {
+class VocabularyModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Page Fragment"
+        value = "This is reading Fragment"
     }
     val text: LiveData<String> = _text
-
 }
