+++ app/src/main/java/com/example/projectdva232v1/ui/Writing/WritingModel.kt
@@ -1,10 +1,10 @@
-package com.example.projectdva232v1.ui.ListeningActivity
+package com.example.projectdva232v1.ui.Writing
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ListeningModel : ViewModel() {
+class WritingModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is reading Fragment"
